Updated converter.py to support different output methods for each type of output.
This helps for outputting excel files where the output file
should be opened in binary mode, whereas JSON and other file types
should be opened as a text file. Additionally, I clarified the errors
for when the the data type of the input file could not be automatically
detected.

I also added in the test-data directory, to provide examples in the
README that can be replicated by the user. Updated requirements.txt
to use more recent versions of packages.